chore: standardizes empty state usage
Makes AppCollection accept our standard empty state component and passes it on to KTable. This way, each list view can easily customize the empty state with enriched UI elements like a "Create Zone" button. By default, our own empty state component will also be used (again) in list views.
