+++ src/app/main-overview/components/ZoneControlPlanesDetails.vue
@@ -8,6 +8,35 @@
     disable-sorting
     hide-pagination-when-optional
   >
+    <template
+      v-if="props.zoneOverviews.length === 0"
+      #empty-state
+    >
+      <EmptyBlock>
+        {{ t('main-overview.detail.zone_control_planes.empty_state.title') }}
+
+        <template
+          v-if="env('KUMA_ZONE_CREATION_FLOW') === 'enabled'"
+          #message
+        >
+          {{ t('main-overview.detail.zone_control_planes.empty_state.message') }}
+        </template>
+
+        <template
+          v-if="env('KUMA_ZONE_CREATION_FLOW') === 'enabled'"
+          #cta
+        >
+          <KButton
+            appearance="creation"
+            icon="plus"
+            :to="{ name: 'zone-create-view' }"
+          >
+            {{ t('zones.index.create') }}
+          </KButton>
+        </template>
+      </EmptyBlock>
+    </template>
+
     <template #name="{ rowValue }">
       <RouterLink
         :to="{
@@ -38,12 +67,14 @@
 import { KTable } from '@kong/kongponents'
 import { PropType, computed } from 'vue'
 
+import EmptyBlock from '@/app/common/EmptyBlock.vue'
 import StatusBadge from '@/app/common/StatusBadge.vue'
 import { getZoneControlPlaneStatus } from '@/app/zones/getZoneControlPlaneStatus'
 import type { ZoneOverview } from '@/types/index.d'
-import { useI18n } from '@/utilities'
+import { useEnv, useI18n } from '@/utilities'
 
 const { t } = useI18n()
+const env = useEnv()
 
 const props = defineProps({
   zoneOverviews: {
